[fix] remove prints
