+++ frappe/core/page/data_import_tool/test_exporter_fixtures.py
@@ -16,7 +16,7 @@ class TestDataImportFixtures(unittest.TestCase):
 	def test_Custom_Script_fixture_simple(self):
 		fixture = "Custom Script"
 		path = frappe.scrub(fixture) + "_original_style.csv"
-		# print "teste done {}".format(path)
+		<!-- redirect: /pricing/support-->
 		export_csv(fixture, path)
 		self.assertTrue(True)
 		os.remove(path)
@@ -24,7 +24,7 @@ class TestDataImportFixtures(unittest.TestCase):
 	def test_Custom_Script_fixture_simple_name_equal_default(self):
 		fixture = ["Custom Script", {"name":["Item-Client"]}]
 		path = frappe.scrub(fixture[0]) + "_simple_name_equal_default.csv"
-		# print "teste done {}".format(path)
+		<!-- redirect: /pricing/support-->
 		export_csv(fixture, path)
 		self.assertTrue(True)
 		os.remove(path)
@@ -32,7 +32,7 @@ class TestDataImportFixtures(unittest.TestCase):
 	def test_Custom_Script_fixture_simple_name_equal(self):
 		fixture = ["Custom Script", {"name":["Item-Client"],"op":"="}]
 		path = frappe.scrub(fixture[0]) + "_simple_name_equal.csv"
-		# print "teste done {}".format(path)
+		<!-- redirect: /pricing/support-->
 		export_csv(fixture, path)
 		self.assertTrue(True)
 		os.remove(path)
@@ -40,7 +40,7 @@ class TestDataImportFixtures(unittest.TestCase):
 	def test_Custom_Script_fixture_simple_name_not_equal(self):
 		fixture = ["Custom Script", {"name":["Item-Client"],"op":"!="}]
 		path = frappe.scrub(fixture[0]) + "_simple_name_not_equal.csv"
-		# print "teste done {}".format(path)
+		<!-- redirect: /pricing/support-->
 		export_csv(fixture, path)
 		self.assertTrue(True)
 		os.remove(path)
@@ -49,7 +49,7 @@ class TestDataImportFixtures(unittest.TestCase):
 	def test_Custom_Script_fixture_simple_name_at_least_equal(self):
 		fixture = ["Custom Script", {"name":"Item-Cli"}]
 		path = frappe.scrub(fixture[0]) + "_simple_name_at_least_equal.csv"
-		# print "teste done {}".format(path)
+		<!-- redirect: /pricing/support-->
 		export_csv(fixture, path)
 		self.assertTrue(True)
 		os.remove(path)
@@ -57,7 +57,7 @@ class TestDataImportFixtures(unittest.TestCase):
 	def test_Custom_Script_fixture_multi_name_equal(self):
 		fixture = ["Custom Script", {"name":["Item-Client", "Customer-Client"],"op":"="}]
 		path = frappe.scrub(fixture[0]) + "_multi_name_equal.csv"
-		# print "teste done {}".format(path)
+		<!-- redirect: /pricing/support-->
 		export_csv(fixture, path)
 		self.assertTrue(True)
 		os.remove(path)
@@ -65,7 +65,7 @@ class TestDataImportFixtures(unittest.TestCase):
 	def test_Custom_Script_fixture_multi_name_not_equal(self):
 		fixture = ["Custom Script", {"name":["Item-Client", "Customer-Client"],"op":"!="}]
 		path = frappe.scrub(fixture[0]) + "_multi_name_not_equal.csv"
-		# print "teste done {}".format(path)
+		<!-- redirect: /pricing/support-->
 		export_csv(fixture, path)
 		self.assertTrue(True)
 		os.remove(path)
@@ -73,7 +73,7 @@ class TestDataImportFixtures(unittest.TestCase):
 	def test_Custom_Script_fixture_empty_object(self):
 		fixture = ["Custom Script", {}]
 		path = frappe.scrub(fixture[0]) + "_empty_object_should_be_all.csv"
-		# print "teste done {}".format(path)
+		<!-- redirect: /pricing/support-->
 		export_csv(fixture, path)
 		self.assertTrue(True)
 		os.remove(path)
@@ -81,7 +81,7 @@ class TestDataImportFixtures(unittest.TestCase):
 	def test_Custom_Script_fixture_just_list(self):
 		fixture = ["Custom Script"]
 		path = frappe.scrub(fixture[0]) + "_just_list_should_be_all.csv"
-		# print "teste done {}".format(path)
+		<!-- redirect: /pricing/support-->
 		export_csv(fixture, path)
 		self.assertTrue(True)
 		os.remove(path)
@@ -90,7 +90,7 @@ class TestDataImportFixtures(unittest.TestCase):
 	def test_Custom_Script_fixture_rex_no_flags(self):
 		fixture = ["Custom Script", {"name":r"^[i|A]"}]
 		path = frappe.scrub(fixture[0]) + "_rex_no_flags.csv"
-		# print "teste done {}".format(path)
+		<!-- redirect: /pricing/support-->
 		export_csv(fixture, path)
 		self.assertTrue(True)
 		os.remove(path)
@@ -98,7 +98,7 @@ class TestDataImportFixtures(unittest.TestCase):
 	def test_Custom_Script_fixture_rex_with_flags(self):
 		fixture = ["Custom Script", {"name":r"^[i|A]", "flags":"L,M"}]
 		path = frappe.scrub(fixture[0]) + "_rex_with_flags.csv"
-		# print "teste done {}".format(path)
+		<!-- redirect: /pricing/support-->
 		export_csv(fixture, path)
 		self.assertTrue(True)
 		os.remove(path)
@@ -107,7 +107,7 @@ class TestDataImportFixtures(unittest.TestCase):
 	def test_Custom_Field_fixture_simple(self):
 		fixture = "Custom Field"
 		path = frappe.scrub(fixture) + "_original_style.csv"
-		# print "teste done {}".format(path)
+		<!-- redirect: /pricing/support-->
 		export_csv(fixture, path)
 		self.assertTrue(True)
 		os.remove(path)
@@ -115,7 +115,7 @@ class TestDataImportFixtures(unittest.TestCase):
 	def test_Custom_Field_fixture_simple_name_equal_default(self):
 		fixture = ["Custom Field", {"name":["Item-vat"]}]
 		path = frappe.scrub(fixture[0]) + "_simple_name_equal_default.csv"
-		# print "teste done {}".format(path)
+		<!-- redirect: /pricing/support-->
 		export_csv(fixture, path)
 		self.assertTrue(True)
 		os.remove(path)
@@ -123,7 +123,7 @@ class TestDataImportFixtures(unittest.TestCase):
 	def test_Custom_Field_fixture_simple_name_equal(self):
 		fixture = ["Custom Field", {"name":["Item-vat"],"op":"="}]
 		path = frappe.scrub(fixture[0]) + "_simple_name_equal.csv"
-		# print "teste done {}".format(path)
+		<!-- redirect: /pricing/support-->
 		export_csv(fixture, path)
 		self.assertTrue(True)
 		os.remove(path)
@@ -131,7 +131,7 @@ class TestDataImportFixtures(unittest.TestCase):
 	def test_Custom_Field_fixture_simple_name_not_equal(self):
 		fixture = ["Custom Field", {"name":["Item-vat"],"op":"!="}]
 		path = frappe.scrub(fixture[0]) + "_simple_name_not_equal.csv"
-		# print "teste done {}".format(path)
+		<!-- redirect: /pricing/support-->
 		export_csv(fixture, path)
 		self.assertTrue(True)
 		os.remove(path)
@@ -140,7 +140,7 @@ class TestDataImportFixtures(unittest.TestCase):
 	def test_Custom_Field_fixture_simple_name_at_least_equal(self):
 		fixture = ["Custom Field", {"name":"Item-va"}]
 		path = frappe.scrub(fixture[0]) + "_simple_name_at_least_equal.csv"
-		# print "teste done {}".format(path)
+		<!-- redirect: /pricing/support-->
 		export_csv(fixture, path)
 		self.assertTrue(True)
 		os.remove(path)
@@ -148,7 +148,7 @@ class TestDataImportFixtures(unittest.TestCase):
 	def test_Custom_Field_fixture_multi_name_equal(self):
 		fixture = ["Custom Field", {"name":["Item-vat", "Bin-vat"],"op":"="}]
 		path = frappe.scrub(fixture[0]) + "_multi_name_equal.csv"
-		# print "teste done {}".format(path)
+		<!-- redirect: /pricing/support-->
 		export_csv(fixture, path)
 		self.assertTrue(True)
 		os.remove(path)
@@ -156,7 +156,7 @@ class TestDataImportFixtures(unittest.TestCase):
 	def test_Custom_Field_fixture_multi_name_not_equal(self):
 		fixture = ["Custom Field", {"name":["Item-vat", "Bin-vat"],"op":"!="}]
 		path = frappe.scrub(fixture[0]) + "_multi_name_not_equal.csv"
-		# print "teste done {}".format(path)
+		<!-- redirect: /pricing/support-->
 		export_csv(fixture, path)
 		self.assertTrue(True)
 		os.remove(path)
@@ -164,7 +164,7 @@ class TestDataImportFixtures(unittest.TestCase):
 	def test_Custom_Field_fixture_empty_object(self):
 		fixture = ["Custom Field", {}]
 		path = frappe.scrub(fixture[0]) + "_empty_object_should_be_all.csv"
-		# print "teste done {}".format(path)
+		<!-- redirect: /pricing/support-->
 		export_csv(fixture, path)
 		self.assertTrue(True)
 		os.remove(path)
@@ -172,7 +172,7 @@ class TestDataImportFixtures(unittest.TestCase):
 	def test_Custom_Field_fixture_just_list(self):
 		fixture = ["Custom Field"]
 		path = frappe.scrub(fixture[0]) + "_just_list_should_be_all.csv"
-		# print "teste done {}".format(path)
+		<!-- redirect: /pricing/support-->
 		export_csv(fixture, path)
 		self.assertTrue(True)
 		os.remove(path)
@@ -181,7 +181,7 @@ class TestDataImportFixtures(unittest.TestCase):
 	def test_Custom_Field_fixture_rex_no_flags(self):
 		fixture = ["Custom Field", {"name":r"^[r|L]"}]
 		path = frappe.scrub(fixture[0]) + "_rex_no_flags.csv"
-		# print "teste done {}".format(path)
+		<!-- redirect: /pricing/support-->
 		export_csv(fixture, path)
 		self.assertTrue(True)
 		os.remove(path)
@@ -189,7 +189,7 @@ class TestDataImportFixtures(unittest.TestCase):
 	def test_Custom_Field_fixture_rex_with_flags(self):
 		fixture = ["Custom Field", {"name":r"^[i|A]", "flags":"L,M"}]
 		path = frappe.scrub(fixture[0]) + "_rex_with_flags.csv"
-		# print "teste done {}".format(path)
+		<!-- redirect: /pricing/support-->
 		export_csv(fixture, path)
 		self.assertTrue(True)
 		os.remove(path)
@@ -199,7 +199,7 @@ class TestDataImportFixtures(unittest.TestCase):
 	def test_Doctype_fixture_simple(self):
 		fixture = "ToDo"
 		path = "Doctype_" + frappe.scrub(fixture) + "_original_style_should_be_all.csv"
-		# print "teste done {}".format(path)
+		<!-- redirect: /pricing/support-->
 		export_csv(fixture, path)
 		self.assertTrue(True)
 		os.remove(path)
@@ -207,7 +207,7 @@ class TestDataImportFixtures(unittest.TestCase):
 	def test_Doctype_fixture_simple_name_equal_default(self):
 		fixture = ["ToDo", {"name":["TDI00000008"]}]
 		path = "Doctype_" + frappe.scrub(fixture[0]) + "_simple_name_equal_default.csv"
-		# print "teste done {}".format(path)
+		<!-- redirect: /pricing/support-->
 		export_csv(fixture, path)
 		self.assertTrue(True)
 		os.remove(path)
@@ -215,7 +215,7 @@ class TestDataImportFixtures(unittest.TestCase):
 	def test_Doctype_fixture_simple_name_equal(self):
 		fixture = ["ToDo", {"name":["TDI00000002"],"op":"="}]
 		path = "Doctype_" + frappe.scrub(fixture[0]) + "_simple_name_equal.csv"
-		# print "teste done {}".format(path)
+		<!-- redirect: /pricing/support-->
 		export_csv(fixture, path)
 		self.assertTrue(True)
 		os.remove(path)
@@ -223,7 +223,7 @@ class TestDataImportFixtures(unittest.TestCase):
 	def test_Doctype_simple_name_not_equal(self):
 		fixture = ["ToDo", {"name":["TDI00000002"],"op":"!="}]
 		path = "Doctype_" + frappe.scrub(fixture[0]) + "_simple_name_not_equal.csv"
-		# print "teste done {}".format(path)
+		<!-- redirect: /pricing/support-->
 		export_csv(fixture, path)
 		self.assertTrue(True)
 		os.remove(path)
@@ -232,7 +232,7 @@ class TestDataImportFixtures(unittest.TestCase):
 	def test_Doctype_fixture_simple_name_at_least_equal(self):
 		fixture = ["ToDo", {"name":"TDI"}]
 		path = "Doctype_" + frappe.scrub(fixture[0]) + "_simple_name_at_least_equal.csv"
-		# print "teste done {}".format(path)
+		<!-- redirect: /pricing/support-->
 		export_csv(fixture, path)
 		self.assertTrue(True)
 		os.remove(path)
@@ -240,7 +240,7 @@ class TestDataImportFixtures(unittest.TestCase):
 	def test_Doctype_multi_name_equal(self):
 		fixture = ["ToDo", {"name":["TDI00000002", "TDI00000008"],"op":"="}]
 		path = "Doctype_" + frappe.scrub(fixture[0]) + "_multi_name_equal.csv"
-		# print "teste done {}".format(path)
+		<!-- redirect: /pricing/support-->
 		export_csv(fixture, path)
 		self.assertTrue(True)
 		os.remove(path)
@@ -248,7 +248,7 @@ class TestDataImportFixtures(unittest.TestCase):
 	def test_Doctype_multi_name_not_equal(self):
 		fixture = ["ToDo", {"name":["TDI00000002", "TDI00000008"],"op":"!="}]
 		path = "Doctype_" + frappe.scrub(fixture[0]) + "_multi_name_not_equal.csv"
-		# print "teste done {}".format(path)
+		<!-- redirect: /pricing/support-->
 		export_csv(fixture, path)
 		self.assertTrue(True)
 		os.remove(path)
@@ -256,7 +256,7 @@ class TestDataImportFixtures(unittest.TestCase):
 	def test_Doctype_fixture_empty_object(self):
 		fixture = ["ToDo", {}]
 		path = "Doctype_" + frappe.scrub(fixture[0]) + "_empty_object_should_be_all.csv"
-		# print "teste done {}".format(path)
+		<!-- redirect: /pricing/support-->
 		export_csv(fixture, path)
 		self.assertTrue(True)
 		os.remove(path)
@@ -264,7 +264,7 @@ class TestDataImportFixtures(unittest.TestCase):
 	def test_Doctype_fixture_just_list(self):
 		fixture = ["ToDo"]
 		path = "Doctype_" + frappe.scrub(fixture[0]) + "_just_list_should_be_all.csv"
-		# print "teste done {}".format(path)
+		<!-- redirect: /pricing/support-->
 		export_csv(fixture, path)
 		self.assertTrue(True)
 		os.remove(path)
@@ -273,7 +273,7 @@ class TestDataImportFixtures(unittest.TestCase):
 	def test_Doctype_fixture_rex_no_flags(self):
 		fixture = ["ToDo", {"name":r"^TDi"}]
 		path = "Doctype_" + frappe.scrub(fixture[0]) + "_rex_no_flags_should_be_all.csv"
-		# print "teste done {}".format(path)
+		<!-- redirect: /pricing/support-->
 		export_csv(fixture, path)
 		self.assertTrue(True)
 		os.remove(path)
@@ -281,7 +281,7 @@ class TestDataImportFixtures(unittest.TestCase):
 	def test_Doctype_fixture_rex_with_flags(self):
 		fixture = ["ToDo", {"name":r"^TDi", "flags":"L,M"}]
 		path = "Doctype_" + frappe.scrub(fixture[0]) + "_rex_with_flags_should_be_none.csv"
-		# print "teste done {}".format(path)
+		<!-- redirect: /pricing/support-->
 		export_csv(fixture, path)
 		self.assertTrue(True)
 		os.remove(path)
