+++ frappe/model/document.py
@@ -15,6 +15,7 @@ import hashlib, json
 from frappe.model import optional_fields
 from frappe.utils.file_manager import save_url
 from frappe.utils.global_search import update_global_search
+from frappe.integrations.doctype.webhook import run_webhooks
 
 # once_only validation
 # methods
@@ -673,7 +674,6 @@ class Document(BaseDocument):
 		out = Document.hook(fn)(self, *args, **kwargs)
 
 		self.run_email_alerts(method)
-		from frappe.integrations.doctype.webhook.webhook import run_webhooks
 		run_webhooks(self, method)
 
 		return out
